Full state history node 1370

+++ db/db_bolt.go
@@ -8,6 +8,9 @@ package db
 
 import (
 	"context"
+	"encoding/binary"
+	"strings"
+	
 	"github.com/pkg/errors"
 	bolt "go.etcd.io/bbolt"
 
@@ -16,6 +19,9 @@ import (
 )
 
 const fileMode = 0600
+
+// ContractKVNameSpace for ignore delete
+var ContractKVNameSpace = "Contract"
 
 // boltDB is KVStore implementation based bolt DB
 type boltDB struct {
@@ -67,6 +73,9 @@ func (b *boltDB) Put(namespace string, key, value []byte) (err error) {
 		err = errors.Wrap(ErrIO, err.Error())
 	}
 	return err
+}
+func (b *boltDB) DB() interface{} {
+	return b.db
 }
 
 // Get retrieves a record
@@ -156,12 +165,14 @@ func (b *boltDB) Commit(batch KVStoreBatch) (err error) {
 						return errors.Wrapf(err, write.errorFormat, write.errorArgs)
 					}
 				} else if write.writeType == Delete {
-					bucket := tx.Bucket([]byte(write.namespace))
-					if bucket == nil {
-						continue
-					}
-					if err := bucket.Delete(write.key); err != nil {
-						return errors.Wrapf(err, write.errorFormat, write.errorArgs)
+					if !strings.EqualFold(write.namespace, ContractKVNameSpace) {
+						bucket := tx.Bucket([]byte(write.namespace))
+						if bucket == nil {
+							continue
+						}
+						if err := bucket.Delete(write.key); err != nil {
+							return errors.Wrapf(err, write.errorFormat, write.errorArgs)
+						}
 					}
 				}
 			}
@@ -218,6 +229,29 @@ func (b *boltDB) CreateCountingIndexNX(name []byte) (CountingIndex, error) {
 		return nil, err
 	}
 	return NewCountingIndex(b.db, b.config.NumRetries, name, size)
+}
+
+// SaveDeletedTrieNode help save trie node that will be deleted in this block
+func (b *boltDB) SaveDeletedTrieNode(batch KVStoreBatch, hei uint64, trieNodeNameSpace string, trieNodeKeyPrefix []byte) (heightToKeyCache KVStoreBatch, err error) {
+	heightToKeyCache = NewCachedBatch()
+	heightBytes := make([]byte, 8)
+	binary.BigEndian.PutUint64(heightBytes, hei)
+	err = b.db.Update(func(tx *bolt.Tx) error {
+		for i := 0; i < batch.Size(); i++ {
+			write, err := batch.Entry(i)
+			if err != nil {
+				return err
+			}
+			// only save trie node in evm's name space
+			if (write.writeType == Delete) && (strings.EqualFold(write.namespace, ContractKVNameSpace)) {
+				heightTo := append(trieNodeKeyPrefix, heightBytes...)
+				heightTo = append(heightTo, write.key...)
+				heightToKeyCache.Put(trieNodeNameSpace, heightTo, []byte(""), write.errorFormat, write.errorArgs)
+			}
+		}
+		return nil
+	})
+	return
 }
 
 //======================================
